% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_comparison_scatter_plot_single.R
\name{algo_comparison_scatter_plot_single}
\alias{algo_comparison_scatter_plot_single}
\title{Scatter Plot for comparing CRISPR screen algorithms or metrics}
\usage{
algo_comparison_scatter_plot_single(item1_file, item2_file, join_on,
  extract1, extract2, main_title, sub_title, fraction_text = 0.01,
  regression_line = F, col_based_on = 1, column_point, alpha = 0.1,
  printPlot = T, savePlot = F, saveDir = NULL, fileType = c("pdf",
  "png", "tiff"))
}
\arguments{
\item{item1_file}{A full filename path to  dataset one}

\item{item2_file}{A full filename path to dataset two}

\item{join_on}{What do you want to join your two datasets on to make a comparison. i.e. 'GENE'}

\item{extract1}{Which column of values to extract from dataset1. i.e. 'jacks_w'}

\item{extract2}{Which column of values to extract from dataset2. i.e. 'BF'}

\item{main_title}{Main title of plot. i.e. 'BAGEL Bayes factor score vs JACKS essentiality score'}

\item{sub_title}{Dataset name for subtitle and savename i.e. "CIS_MO_T18_tkoV2_2017_04"}

\item{fraction_text}{fraction of outlying point identifiers taken from join_on column to plot on graph. lower value prints less labels}

\item{regression_line}{add a regression line to plot}

\item{col_based_on}{color points based on which dataset, 1 = dataset1, 2 = dataset2}

\item{column_point}{Which column(s) to use to identifier different points. Ex. c('jacks_neg_fdr','jacks_pos_fdr'), should be a numeric colour}

\item{alpha}{Threshold for what to colour different points as based on. If using FDR column, then FDR < 0.1 --> will have different colour from rest}

\item{savePlot}{Option to save plot to disk}

\item{saveDir}{directory you wish to save plots to}

\item{fileType}{type of plots you want to save}

\item{printPlt}{Print plot to R graphics}
}
\value{
A plot comparing the same dataset analysed for different metrics or algorithms, and returns scatter plot object
}
\description{
This script can used to compare two different datasets of the same nature.
Meaning they can be from the same treatment/condition and have something that
they can be compared on, i.e. Genes.
}
\details{
This will create a scatter plot comparing two metrics from the two datasets.
For example, comparing JACKS w essentiality score vs BAGELS baye factor score.
}
\author{
Justin Sing, \url{https://github.com/singjc}
}
