% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_Counts_PreProcessing.R
\name{Raw_Counts_PreProcessing}
\alias{Raw_Counts_PreProcessing}
\title{Raw Counts PreProcessing}
\usage{
Raw_Counts_PreProcessing(df, meta_cols = c(1, 2),
  col_with_replicates = NULL, nreplicates = 2, n = 100, std = 100,
  t_stat_rep_alpha = 5e-07, data_process_partition = 10,
  consistency_cutoff = 0.3, printPlots = F, filename = "",
  saveDir = NULL, savePlots = F, nrdiPlotstoSave = 1500)
}
\arguments{
\item{df}{data.frame of count data containing the sgRNA, Genes columns and 2 replicate columns for each condition}

\item{meta_cols}{the two columns that contain the information for the sgRNA ids and the gene ids}

\item{col_with_replicates}{the columns in df that contain replicates. Replicate samples need to be right next to each other for this script to work}

\item{nreplicates}{the number of replicates per sample}

\item{n}{number of quantiles/vectors to make for rnorm distribution}

\item{std}{the standard deviation to create a rnorm distribution}

\item{t_stat_rep_alpha}{the p-value cutt-off for t-test, to decide whether the replicate distributions are significantly different or not}

\item{data_process_partition}{the amount you want to split your data into, to make processing quicker, especially when using mclapply}

\item{consistency_cutoff}{the ratio of acceptable inconsistent sgRNAs to allow}

\item{printPlots}{option to print plots}

\item{filename}{File name to save filtered Results as}

\item{saveDir}{Directory to save results. Ex. /User/Documents}

\item{savePlots}{option to save plots. T or F}

\item{nrdiPlotstoSave}{(Not Working Yet) The number of every nth  replicate distribution intersection plot to save}
}
\value{
Writes filtered counts matrix to dataframe to disk, and sgRNA counts removed dataframe to disk, and returns both stored in a list variable
}
\description{
This script is a callable function script used to visualize raw counts data
to check for any outliers when comparing different guides of the same gene.
There is some natural variation one would expect, but if there is a large
variation for one guide compared to the next, then this will cause a bias
towards this gene being over expressed or significant. The best case scenario
is to remove these outlying guides.
}
\author{
Justin Sing, \url{https://github.com/singjc}
}
